Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
pydantic[email]
cloudinary
python-multipart
bcrypt==4.0.1
passlib[bcrypt]
